+++ verticapy/core/_display.py
@@ -14,7 +14,7 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-import shutil
+import shutil, html
 
 
 def print_table(
